Portfolio evaluation 1. Implemented: periodically started service that calculates portfolio valuation at the end of the previous day. 2. Added to settings fragment: Selector at what time user wants (if so) to receive daily notification about portfolio state.

+++ app/src/main/java/dev/kokorev/cryptoview/utils/PortfolioInteractor.kt
@@ -7,7 +7,9 @@ import android.view.View
 import androidx.core.content.ContextCompat.getString
 import androidx.core.widget.addTextChangedListener
 import com.coinpaprika.apiclient.entity.CoinDetailsEntity
-import com.coinpaprika.apiclient.entity.PortfolioCoinDB
+import com.coinpaprika.apiclient.entity.PortfolioEvaluationDB
+import com.coinpaprika.apiclient.entity.PortfolioPositionDB
+import com.coinpaprika.apiclient.entity.PortfolioTransactionDB
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.snackbar.Snackbar
 import dev.kokorev.cryptoview.App
@@ -16,6 +18,7 @@ import dev.kokorev.cryptoview.databinding.AlertViewChangePositionBinding
 import dev.kokorev.cryptoview.databinding.AlertViewOpenPositionBinding
 import dev.kokorev.cryptoview.domain.Repository
 import dev.kokorev.cryptoview.views.fragments.toEditable
+import java.time.Instant
 import javax.inject.Inject
 
 class PortfolioInteractor(val view: View, val autoDisposable: AutoDisposable) {
@@ -91,8 +94,10 @@ class PortfolioInteractor(val view: View, val autoDisposable: AutoDisposable) {
                         }
                         .show()
                 } else {
-                    val portfolioCoinDB: PortfolioCoinDB = Converter.createPortfolioCoin(coin, price, qty)
-                    repository.savePortfolioPosition(portfolioCoinDB)
+                    val portfolioPositionDB: PortfolioPositionDB = Converter.createPortfolioPosition(coin, price, qty)
+                    
+                    performTransaction(portfolioPositionDB, price, qty)
+                    
                     MaterialAlertDialogBuilder(context, R.style.CVDialogStyle)
                         .setTitle(getString(context, R.string.portfolio_operation))
                         .setMessage(getString(context, R.string.position_created))
@@ -100,7 +105,7 @@ class PortfolioInteractor(val view: View, val autoDisposable: AutoDisposable) {
                             dialogEmptyInput.cancel()
                         }
                         .show()
-                    Log.d(this.javaClass.simpleName, "askUserQtyToOpenPosition added to portfolio: ${portfolioCoinDB}")
+                    Log.d(this.javaClass.simpleName, "askUserQtyToOpenPosition added to portfolio: ${portfolioPositionDB}")
                 }
                 Log.d(this.javaClass.simpleName, "askUserQtyToOpenPosition user input: ${qty}")
                 dialog.cancel()
@@ -112,23 +117,21 @@ class PortfolioInteractor(val view: View, val autoDisposable: AutoDisposable) {
         return
     }
 
-
-
-
-    fun changePosition(portfolioCoinDB: PortfolioCoinDB) {
-        repository.getCPTickerById(portfolioCoinDB.coinPaprikaId)
+    
+    fun changePosition(portfolioPositionDB: PortfolioPositionDB) {
+        repository.getCPTickerById(portfolioPositionDB.coinPaprikaId)
             .doOnSuccess { ticker ->
                 val price = ticker.price ?: 0.0
                 Log.d(
                     this.javaClass.simpleName,
-                    "Changing position coin: ${portfolioCoinDB.symbol}, price: ${price}"
+                    "Changing position coin: ${portfolioPositionDB.symbol}, price: ${price}"
                 )
-                askUserQtyToChangePosition(portfolioCoinDB, price)
+                askUserQtyToChangePosition(portfolioPositionDB, price)
             }
             .doOnComplete {
                 Snackbar.make(
                     view,
-                    getString(context, R.string.price_not_found) + portfolioCoinDB.symbol,
+                    getString(context, R.string.price_not_found) + portfolioPositionDB.symbol,
                     Snackbar.LENGTH_SHORT
                 ).show()
             }
@@ -137,7 +140,7 @@ class PortfolioInteractor(val view: View, val autoDisposable: AutoDisposable) {
     }
 
     fun askUserQtyToChangePosition(
-        position: PortfolioCoinDB,
+        position: PortfolioPositionDB,
         price: Double
     ) {
         if (price == 0.0) return
@@ -184,12 +187,11 @@ class PortfolioInteractor(val view: View, val autoDisposable: AutoDisposable) {
                 dialog.cancel()
             }
             .show()
-
         return
     }
 
     private fun changePosition(
-        position: PortfolioCoinDB,
+        position: PortfolioPositionDB,
         newQty: Double,
         price: Double
     ) {
@@ -205,11 +207,13 @@ class PortfolioInteractor(val view: View, val autoDisposable: AutoDisposable) {
                 dialogEmptyInput.cancel()
             }
             .show()
+        val dealValue = price * (newQty - position.quantity)
         val newPrice =
-            (position.priceOpen * position.quantity + price * (newQty - position.quantity)) / newQty
+            (position.priceOpen * position.quantity + dealValue) / newQty
+        val oldQty = position.quantity
         position.quantity = newQty
         position.priceOpen = newPrice
-        repository.savePortfolioPosition(position)
+        performTransaction(position, price = price, qty = newQty - oldQty)
     }
 
     private fun showNoChanges() {
@@ -228,10 +232,11 @@ class PortfolioInteractor(val view: View, val autoDisposable: AutoDisposable) {
     }
 
     private fun closePosition(
-        position: PortfolioCoinDB,
+        position: PortfolioPositionDB,
         price: Double
     ) {
-        val pnl = position.quantity * (price - position.priceOpen)
+        val oldQty = position.quantity
+        val pnl = oldQty * (price - position.priceOpen)
         val pnlStr = NumbersUtils.formatPriceUSD(pnl)
         MaterialAlertDialogBuilder(context, R.style.CVDialogStyle)
             .setTitle(getString(context, R.string.portfolio_operation))
@@ -245,6 +250,59 @@ class PortfolioInteractor(val view: View, val autoDisposable: AutoDisposable) {
                 dialogEmptyInput.cancel()
             }
             .show()
-        repository.deletePortfolioPosition(position.id)
+        position.quantity = 0.0
+        performTransaction(position, price = price, qty = -oldQty)
+    }
+    
+    
+    private fun performTransaction(position: PortfolioPositionDB, price: Double, qty: Double) {
+        // saving portfolio state
+        if (position.quantity == 0.0) {
+            repository.deletePortfolioPosition(position.id)
+        } else {
+            repository.savePortfolioPosition(position)
+        }
+        // saving inflow/outflow for today's portfolio valuation
+        saveInflowToPortfolioEvaluation(price * qty)
+        // saving the transaction to db
+        saveTransaction(position.coinPaprikaId, price, qty)
+    }
+    
+    private fun saveTransaction(coinPaprikaId: String, price: Double, qty: Double) {
+        val transaction = PortfolioTransactionDB(
+            coinPaprikaId = coinPaprikaId,
+            time = Instant.now(),
+            price = price,
+            quantity = qty,
+        )
+        repository.savePortfolioTransaction(transaction)
+    }
+    
+    private fun saveInflowToPortfolioEvaluation(dealValue: Double) {
+        val today = Instant.now().toLocalDate()
+
+        repository.getPortfolioEvaluationByDate(today)
+            .doOnSuccess {  evaluation ->
+                Log.d(this.javaClass.simpleName, "Portfolio evaluation found for today. Valuation: ${evaluation.valuation}, Inflow: ${evaluation.inflow}")
+                evaluation.inflow = (evaluation.inflow ?: 0.0) + dealValue
+                repository.savePortfolioEvaluation(evaluation)
+            }
+            .doOnComplete {
+                Log.d(this.javaClass.simpleName, "No portfolio evaluation found for today")
+                val portfolioEvaluation = PortfolioEvaluationDB(
+                    date = today,
+                    valuation = null,
+                    inflow = dealValue,
+                    change = null,
+                    percentChange = null,
+                    positions = null,
+                )
+                repository.savePortfolioEvaluation(portfolioEvaluation)
+            }
+            .doOnError {
+                Log.d(this.javaClass.simpleName, "Error getting portfolio evaluation, message: ${it.localizedMessage}, ${it.stackTrace}")
+            }
+            .subscribe()
+            .addTo(autoDisposable)
     }
 }
